Update README.md
I have augmented my professional portfolio with the inclusion of TypeScript and Angular, bolstering my proficiency in these technologies. This enhancement aligns with my ongoing exploration of the synergistic potential inherent in the Asp.net + Angular stack. As a conscientious developer, I continually seek to expand my skill set and knowledge base to deliver exceptional solutions in the ever-evolving world of web development.